refactor(scheduler): extract post-group boundaries (#1918)
