+++ apps/server/api/src/collections/post-groups/post-groups.module.ts
@@ -1,4 +1,6 @@
 import { PostGroupsController } from '@api/collections/post-groups/controllers/post-groups.controller';
+import { PostGroupContractService } from '@api/collections/post-groups/services/post-group-contract.service';
+import { PostGroupPersistenceService } from '@api/collections/post-groups/services/post-group-persistence.service';
 import { PostGroupsService } from '@api/collections/post-groups/services/post-groups.service';
 import { PublishApprovalsModule } from '@api/collections/publish-approvals/publish-approvals.module';
 import { QueuesModule } from '@api/queues/core/queues.module';
@@ -8,6 +10,10 @@ import { forwardRef, Module } from '@nestjs/common';
   controllers: [PostGroupsController],
   exports: [PostGroupsService],
   imports: [forwardRef(() => QueuesModule), PublishApprovalsModule],
-  providers: [PostGroupsService],
+  providers: [
+    PostGroupContractService,
+    PostGroupPersistenceService,
+    PostGroupsService,
+  ],
 })
 export class PostGroupsModule {}
